+++ src/runtime/styling/index.ts
@@ -28,11 +28,30 @@ export interface StyledElements extends StyledIntrinsicFactories {
     elementType: T,
     ...tokens: readonly MaybeThunk<React.CSSProperties | Falsy>[]
   ): React.ReactHTML[T];
+
+  // styled(Component)`...`
+  <P extends {} = {}>(component: (props: P) => JSX.Element): StyledComponent<P>;
+  // styled(Component, {...}) || styled(Component, '...')
+  <P extends {} = {}>(
+    component: (props: P) => JSX.Element,
+    ...tokens: readonly MaybeThunk<React.CSSProperties | Falsy>[]
+  ): (props: P) => JSX.Element;
 }
 
 type Falsy = '' | 0 | -0 | false | null | undefined | void;
 type MaybeArray<T> = T | readonly T[];
 type MaybeThunk<T> = T | ((context: Context) => T);
+
+interface StyledComponent<P extends {} = {}> {
+  (
+    strings: TemplateStringsArray,
+    ...interpolations: readonly MaybeThunk<
+      MaybeArray<React.CSSProperties | string | number | Falsy>
+    >[]
+  ): (props: P) => JSX.Element;
+  (tokens: MaybeThunk<MaybeArray<React.CSSProperties | Falsy>>): (props: P) => JSX.Element;
+  (...tokens: readonly MaybeThunk<React.CSSProperties | Falsy>[]): (props: P) => JSX.Element;
+}
 
 interface StyledHTML<T extends keyof React.ReactHTML> {
   (
@@ -45,9 +64,27 @@ interface StyledHTML<T extends keyof React.ReactHTML> {
   (...tokens: readonly MaybeThunk<React.CSSProperties | Falsy>[]): React.ReactHTML[T];
 }
 
-function styledHtml<T extends keyof React.ReactHTML>(elementType: T): StyledHTML<T> {
+function styledComponent<P extends {} = {}>(
+  component: (props: P) => JSX.Element,
+  parentDirective?: Directive<CSSRules>
+): StyledComponent<P> {
   return function boundStyled(...args: any[]) {
-    const directive = css(...args);
+    const directive = css(parentDirective, css(...args));
+
+    // Can't figure out the 100% accurate typing here so we're going to bail out with some
+    // anys.
+    return function boundStyledElement(props?: any): any {
+      return renderStyled(component, props, directive);
+    };
+  };
+}
+
+function styledHtml<T extends keyof React.ReactHTML>(
+  elementType: T,
+  parentDirective?: Directive<CSSRules>
+): StyledHTML<T> {
+  return function boundStyled(...args: any[]) {
+    const directive = css(parentDirective, css(...args));
 
     // Can't figure out the 100% accurate typing here so we're going to bail out with some
     // anys.
@@ -58,7 +95,7 @@ function styledHtml<T extends keyof React.ReactHTML>(elementType: T): StyledHTML
 }
 
 function renderStyled(
-  elementType: keyof React.ReactHTML,
+  elementType: keyof React.ReactHTML | ((props: any) => JSX.Element),
   { className, ...props }: any = {},
   directive: Directive<CSSRules>
 ) {
@@ -67,19 +104,24 @@ function renderStyled(
 
   return React.createElement(elementType, {
     ...props,
-    className: `${className}${className ? ' ' : ''}${styledClassName}`,
+    className: `${className ? `${className} ` : ''}${styledClassName}`,
   });
 }
 
-export const styled: StyledElements = new Proxy(Object.create(null), {
-  apply(_target: any, _this: any, elementType: keyof React.ReactHTML, ...args: any) {
+export const styled: StyledElements = new Proxy(renderStyled, {
+  apply(
+    _target: any,
+    _this: any,
+    [elementType, ...args]: [
+      elementType: keyof React.ReactHTML | ((props: any) => JSX.Element),
+      ...args: any[]
+    ]
+  ) {
     const directive = css(...args);
 
-    // Can't figure out the 100% accurate typing here so we're going to bail out with some
-    // anys.
-    return function boundStyledElement({ className, ...props }: any = {}): any {
-      return renderStyled(elementType, props, directive);
-    };
+    return typeof elementType === 'string'
+      ? styledHtml(elementType, directive)
+      : styledComponent(elementType, directive);
   },
   get(_t, elementType: keyof React.ReactHTML) {
     return styledHtml(elementType);