+++ src/worker/mdxCompiler.ts
@@ -1,25 +1,37 @@
-import { BUNDLED_THEMES, getHighlighter, loadTheme } from '@antfu/shiki';
-import { htmlEscape } from 'escape-goat';
 import grayMatter from 'gray-matter';
-import type { Element, Text } from 'hast';
-import LRUCache from 'lru-cache';
 import * as Path from 'path';
+import * as Querystring from 'querystring';
 //@ts-ignore
 import remarkAutolinkHeadings from 'remark-autolink-headings';
 import remarkGfm from 'remark-gfm';
 //@ts-ignore
 import remarkSlug from 'remark-slug';
-import visit, { SKIP } from 'unist-util-visit';
+import type unified from 'unified';
+import type unist from 'unist';
+import visit from 'unist-util-visit';
 import { compile } from 'xdm';
+import { createSnippetData } from '../runtime/internal/components/snippetParser';
 
-const cache = new LRUCache<string, Element>({
-  max: 1024 * 1024 * 32, // 32 mb
-  length(value, key) {
-    return JSON.stringify(value).length + (key ? key.length : 0);
-  },
-});
+interface FlowElementAttributeExpression {
+  type: 'mdxJsxAttributeValueExpression';
+  data: {
+    estree: any;
+  };
+}
 
-export default async function compileMdx([path, contents]: [path: string, contents: string]) {
+type FlowElementAttributeValue = string | FlowElementAttributeExpression;
+
+interface FlowElementAttribute {
+  type: 'mdxJsxAttribute';
+  name: string;
+  value: FlowElementAttributeValue;
+}
+
+export default function handleWorkerRequest([path, contents]: [string, string]) {
+  return compileMdx(path, contents);
+}
+
+export async function compileMdx(path: string, contents: string) {
   const parsed = grayMatter(contents, {
     excerpt: true,
   });
@@ -33,12 +45,19 @@ export default async function compileMdx([path, contents]: [path: string, conten
     );
   }
 
+  const remarkPlugins: unified.PluggableList<unified.Settings> = [
+    remarkSnippets,
+    remarkGfm,
+    remarkAutolinkHeadings,
+    remarkSlug,
+  ];
+
   const mdxJsx = await compile(
     { contents: parsed.content, path },
     {
       jsx: true,
       jsxRuntime: 'classic',
-      remarkPlugins: [remarkCodeBlocksShiki, remarkGfm, remarkAutolinkHeadings, remarkSlug],
+      remarkPlugins,
     }
   );
 
@@ -54,106 +73,187 @@ export default async function compileMdx([path, contents]: [path: string, conten
   return transformed;
 }
 
-const remarkCodeBlocksShiki: import('unified').Plugin = () => {
-  return async function transformer(tree) {
-    // The path is relative to the dist dir where themes will be put in a themes folder
-    const oneDark = await loadTheme(Path.resolve(__dirname, './themes/OneDark.json'));
-    const highlighter = await getHighlighter({ themes: [...BUNDLED_THEMES, oneDark] });
-    const themeOptionRx = /^theme:(.+)$/;
+/**
+ * remarkSnippets is a remark plugin that replaces code blocks with an instance
+ * of the Nostalgie `<CodeSnippet>` element, with an augmented representation
+ * of the original source code injected as a static prop.
+ */
+const remarkSnippets: unified.Plugin = () => {
+  return async function transformer(tree, file) {
+    const codeNodes: unist.Node[] = [];
 
+    // Collect all code blocks w/ a 'lang' attribute
     visit(tree, 'code', (node) => {
       if (!node.lang || !node.value) {
         return;
       }
 
-      const meta = Array.isArray(node.meta) ? node.meta : node.meta ? [node.meta] : [];
-      const defaultTheme: string = 'One Dark Pro';
-      let requestedTheme: string = defaultTheme;
-
-      for (const entry of meta) {
-        const matches = entry.match(themeOptionRx);
-
-        if (matches) {
-          requestedTheme = matches[1];
-        }
-      }
-
-      const cacheKey = JSON.stringify([requestedTheme, node.lang, node.value]);
-      let nodeValue = cache.get(cacheKey);
-
-      if (!nodeValue) {
-        const fgColor = highlighter.getForegroundColor(requestedTheme).toUpperCase();
-        const bgColor = highlighter.getBackgroundColor(requestedTheme).toUpperCase();
-        const tokens = highlighter.codeToThemedTokens(
-          node.value as string,
-          node.lang as string,
-          requestedTheme
-        );
-        const children = tokens.map(
-          (lineTokens, zeroBasedLineNumber): Element => {
-            const children = lineTokens.map((token): Text | Element => {
-              const color = token.color;
-              const content: Text = {
-                type: 'text',
-                // Do not escape the _actual_ content
-                value: token.content,
-              };
-
-              return color && color !== fgColor
-                ? {
-                    type: 'element',
-                    tagName: 'span',
-                    properties: {
-                      style: `color: ${htmlEscape(color)}`,
-                    },
-                    children: [content],
-                  }
-                : content;
-            });
-
-            children.push({
-              type: 'text',
-              value: '\n',
-            });
-
-            return {
-              type: 'element',
-              tagName: 'span',
-              properties: {
-                className: 'codeblock-line',
-                dataLineNumber: zeroBasedLineNumber + 1,
-              },
-              children,
-            };
-          }
-        );
-
-        nodeValue = {
-          type: 'element',
-          tagName: 'pre',
-          properties: {
-            dataLang: htmlEscape(node.lang as string),
-            style: `color: ${htmlEscape(fgColor)};background-color: ${htmlEscape(bgColor)}`,
-          },
-          children: [
-            {
-              type: 'element',
-              tagName: 'code',
-              children,
-            },
-          ],
-        };
-
-        cache.set(cacheKey, nodeValue);
-      }
-
-      const data = node.data ?? (node.data = {});
-
-      node.type = 'element';
-      data.hProperties ??= {};
-      data.hChildren = [nodeValue];
-
-      return SKIP;
+      codeNodes.push(node);
     });
+
+    if (codeNodes.length) {
+      (tree.children as unist.Node[]).unshift({
+        type: 'mdxjsEsm',
+        value: `import { CodeSnippet } from 'nostalgie/internal/components';`,
+        data: {
+          estree: createImportAst(),
+        },
+      });
+
+      await Promise.all(
+        codeNodes.map(async (node) => {
+          const options = Querystring.parse((node.meta as string) || '', ' ', ':');
+          const theme = typeof options.theme === 'string' ? options.theme : undefined;
+          const snippetData = await createSnippetData(node.value as string, {
+            fromPath: file.path,
+            theme,
+            lang: node.lang as string,
+          });
+
+          let emphasizeRanges: Array<[fromLine: number, toLine: number]> | undefined;
+
+          if (options.emphasize) {
+            const rawRanges = Array.isArray(options.emphasize)
+              ? options.emphasize
+              : [options.emphasize];
+
+            emphasizeRanges = rawRanges.map((range) => {
+              // Ranges starts and ends can represented by start:end or start-end.
+              const [start, end] = range.split(/[-:]/, 2);
+
+              return [parseInt(start, 10), end ? parseInt(end, 10) : parseInt(start, 10)];
+            });
+          }
+
+          node.children = [];
+          node.type = 'mdxJsxFlowElement';
+          node.name = 'CodeSnippet';
+          node.attributes = [
+            {
+              name: 'parseResult',
+              type: 'mdxJsxAttribute',
+              value: {
+                type: 'mdxJsxAttributeValueExpression',
+                data: {
+                  estree: createJSONParseAST(JSON.stringify(snippetData)),
+                },
+              },
+            },
+          ] as FlowElementAttribute[];
+
+          if (emphasizeRanges) {
+            (node.attributes as any[]).push({
+              name: 'emphasizeRanges',
+              type: 'mdxJsxAttribute',
+              value: {
+                type: 'mdxJsxAttributeValueExpression',
+                data: {
+                  estree: createRangesArrayAST(emphasizeRanges),
+                },
+              },
+            });
+          }
+        })
+      );
+    }
   };
 };
+
+function createImportAst() {
+  return {
+    type: 'Program',
+    body: [
+      {
+        type: 'ImportDeclaration',
+        specifiers: [
+          {
+            type: 'ImportSpecifier',
+            imported: {
+              type: 'Identifier',
+              name: 'CodeSnippet',
+            },
+            local: {
+              type: 'Identifier',
+              name: 'CodeSnippet',
+            },
+          },
+        ],
+        source: {
+          type: 'Literal',
+          value: 'nostalgie/internal/components',
+          raw: "'nostalgie/internal/components'",
+        },
+      },
+    ],
+    sourceType: 'module',
+    comments: [],
+  };
+}
+
+function createRangesArrayAST(ranges: Array<[start: number, end: number]>) {
+  return {
+    type: 'Program',
+    body: [
+      {
+        type: 'ExpressionStatement',
+        expression: {
+          type: 'ArrayExpression',
+          elements: ranges.map(([start, end]) => ({
+            type: 'ArrayExpression',
+            elements: [
+              {
+                type: 'Literal',
+                value: start,
+                raw: JSON.stringify(start),
+              },
+              {
+                type: 'Literal',
+                value: end,
+                raw: JSON.stringify(end),
+              },
+            ],
+          })),
+        },
+      },
+    ],
+    sourceType: 'module',
+    comments: [],
+  };
+}
+
+function createJSONParseAST(value: string) {
+  return {
+    type: 'Program',
+    body: [
+      {
+        type: 'ExpressionStatement',
+        expression: {
+          type: 'CallExpression',
+          callee: {
+            type: 'MemberExpression',
+            object: {
+              type: 'Identifier',
+              name: 'JSON',
+            },
+            property: {
+              type: 'Identifier',
+              name: 'parse',
+            },
+            computed: false,
+            optional: false,
+          },
+          arguments: [
+            {
+              type: 'Literal',
+              value: value,
+              raw: JSON.stringify(value),
+            },
+          ],
+          optional: false,
+        },
+      },
+    ],
+    sourceType: 'module',
+    comments: [],
+  };
+}